% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_windows.R
\name{Datalist_sliding_w}
\alias{Datalist_sliding_w}
\title{calculate correlation coefficients with ceRNAs of a miRNA by sliding windows.}
\usage{
Datalist_sliding_w(w, dictionary, miRNA_total, mirna_sam, gene_sam)
}
\arguments{
\item{w}{window size of the each calculation of triplets,the default is 10.}

\item{dictionary}{miRNA and the corresponding genes' combinatino in list format, with miRNA name in the first column and the corresponding genes in the second column}

\item{miRNA_total}{miRNA list in a vector format, the order of miRNA_total should be the same as the first column of dictionary}

\item{mirna_sam}{expression data of miRNA in dataframe format, with miRNA's name in rows and sample name in columns}

\item{gene_sam}{expression data of gene in dataframe format, with gene's name in rows and sample name in columns}
}
\value{
Realdata a list of dataframe formats with correlation coefficients of genes within each window for each corresponding miRNA average expression.
}
\description{
This function will calculate correlation coefficients within each window, a sliding mover that contains putative ceRNA triplets composed of a miRNA and several genes, after determing a window size and the number of windows. And the corresponding miRNA expression of each window was the average expression of samples within different windows.
}
\examples{
## Use the internal dataset

data("dictionary", package = "ceRNAmiRNAfun", envir = environment())
data("miRNA_total", package = "ceRNAmiRNAfun", envir = environment())
data("mirna_sam", package = "ceRNAmiRNAfun", envir = environment())
data("gene_sam", package = "ceRNAmiRNAfun", envir = environment())

## use sliding window to calculate correlations
Datalist_sliding_w(w=10,dictionary,miRNA_total,mirna_sam,gene_sam)


}
