% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{A list with information of miRNAs and corresponding genes}
\format{A list with 137 miRNA names in the first column and the corresponding genes in list format in the second column.}
\source{
TCGA lung adenocarcinoma:
 \url{https://tcga-data.nci.nih.gov/docs/publications/tcga/}
}
\usage{
dictionary
}
\value{
list
}
\description{
This list is originally from TCGA lung adenocarcinoma.
}
\keyword{datasets}
