% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gene_sam.R
\docType{data}
\name{gene_sam}
\alias{gene_sam}
\title{gene expression data about lung cancer}
\format{A data frame with 918 rows(genes) and 475 variables:
\describe{
  \item{TCGA.77.A5G6.01A}{sample name, in column}	#'   \item{TCGA.22.4609.01A}{sample name, in column}	#'   \item{TCGA.56.A4BY.01A}{sample name, in column}	#'   \item{TCGA.63.6202.01A}{sample name, in column}	#'   \item{TCGA.66.2786.01A}{sample name, in column}	#'   \item{TCGA.66.2766.01A}{sample name, in column}	#'   \item{TCGA.56.7822.01A}{sample name, in column}	#'   \item{TCGA.39.5011.01A}{sample name, in column}	#'   \item{TCGA.77.8139.01A}{sample name, in column}	#'   \item{TCGA.21.1078.01A}{sample name, in column}	#'   \item{TCGA.22.5474.01A}{sample name, in column}	#'   \item{TCGA.33.4533.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HR.01A}{sample name, in column}	#'   \item{TCGA.33.4586.01A}{sample name, in column}	#'   \item{TCGA.33.4583.01A}{sample name, in column}	#'   \item{TCGA.39.5037.01A}{sample name, in column}	#'   \item{TCGA.60.2714.01A}{sample name, in column}	#'   \item{TCGA.60.2725.01A}{sample name, in column}	#'   \item{TCGA.22.5472.01A}{sample name, in column}	#'   \item{TCGA.22.4591.01A}{sample name, in column}	#'   \item{TCGA.56.8625.01A}{sample name, in column}	#'   \item{TCGA.98.8022.01A}{sample name, in column}	#'   \item{TCGA.85.8276.01A}{sample name, in column}	#'   \item{TCGA.46.3766.01A}{sample name, in column}	#'   \item{TCGA.56.7221.01A}{sample name, in column}	#'   \item{TCGA.60.2709.01A}{sample name, in column}	#'   \item{TCGA.66.2765.01A}{sample name, in column}	#'   \item{TCGA.66.2787.01A}{sample name, in column}	#'   \item{TCGA.90.7964.01A}{sample name, in column}	#'   \item{TCGA.18.3407.01A}{sample name, in column}	#'   \item{TCGA.77.7142.01A}{sample name, in column}	#'   \item{TCGA.63.A5MG.01A}{sample name, in column}	#'   \item{TCGA.34.2600.01A}{sample name, in column}	#'   \item{TCGA.21.5787.01A}{sample name, in column}	#'   \item{TCGA.NK.A7XE.01A}{sample name, in column}	#'   \item{TCGA.94.7943.01A}{sample name, in column}	#'   \item{TCGA.22.4595.01A}{sample name, in column}	#'   \item{TCGA.77.8154.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HK.01A}{sample name, in column}	#'   \item{TCGA.18.3411.01A}{sample name, in column}	#'   \item{TCGA.63.A5MY.01A}{sample name, in column}	#'   \item{TCGA.43.6771.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HG.01A}{sample name, in column}	#'   \item{TCGA.LA.A446.01A}{sample name, in column}	#'   \item{TCGA.96.A4JL.01A}{sample name, in column}	#'   \item{TCGA.98.8021.01A}{sample name, in column}	#'   \item{TCGA.22.4594.01A}{sample name, in column}	#'   \item{TCGA.66.2790.01A}{sample name, in column}	#'   \item{TCGA.85.A50Z.01A}{sample name, in column}	#'   \item{TCGA.56.A62T.01A}{sample name, in column}	#'   \item{TCGA.22.4605.01A}{sample name, in column}	#'   \item{TCGA.66.2791.01A}{sample name, in column}	#'   \item{TCGA.94.A5I4.01A}{sample name, in column}	#'   \item{TCGA.77.A5GF.01A}{sample name, in column}	#'   \item{TCGA.60.2719.01A}{sample name, in column}	#'   \item{TCGA.34.2608.01A}{sample name, in column}	#'   \item{TCGA.22.5482.01A}{sample name, in column}	#'   \item{TCGA.18.3416.01A}{sample name, in column}	#'   \item{TCGA.39.5040.01A}{sample name, in column}	#'   \item{TCGA.39.5024.01A}{sample name, in column}	#'   \item{TCGA.43.2578.01A}{sample name, in column}	#'   \item{TCGA.43.7656.01A}{sample name, in column}	#'   \item{TCGA.66.2793.01A}{sample name, in column}	#'   \item{TCGA.56.7579.01A}{sample name, in column}	#'   \item{TCGA.56.A4BX.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HM.01A}{sample name, in column}	#'   \item{TCGA.58.A46N.01A}{sample name, in column}	#'   \item{TCGA.77.8133.01A}{sample name, in column}	#'   \item{TCGA.77.7140.01A}{sample name, in column}	#'   \item{TCGA.68.A59J.01A}{sample name, in column}	#'   \item{TCGA.77.6843.01A}{sample name, in column}	#'   \item{TCGA.18.3417.01A}{sample name, in column}	#'   \item{TCGA.77.8007.01A}{sample name, in column}	#'   \item{TCGA.18.4086.01A}{sample name, in column}	#'   \item{TCGA.18.3408.01A}{sample name, in column}	#'   \item{TCGA.21.5783.01A}{sample name, in column}	#'   \item{TCGA.60.2720.01A}{sample name, in column}	#'   \item{TCGA.34.7107.01A}{sample name, in column}	#'   \item{TCGA.33.AASJ.01A}{sample name, in column}	#'   \item{TCGA.94.8490.01A}{sample name, in column}	#'   \item{TCGA.77.7139.01A}{sample name, in column}	#'   \item{TCGA.98.A53C.01A}{sample name, in column}	#'   \item{TCGA.98.A53I.01A}{sample name, in column}	#'   \item{TCGA.21.5784.01A}{sample name, in column}	#'   \item{TCGA.56.6546.01A}{sample name, in column}	#'   \item{TCGA.58.8393.01A}{sample name, in column}	#'   \item{TCGA.56.7223.01A}{sample name, in column}	#'   \item{TCGA.33.4582.01A}{sample name, in column}	#'   \item{TCGA.33.4566.01A}{sample name, in column}	#'   \item{TCGA.37.3789.01A}{sample name, in column}	#'   \item{TCGA.77.6845.01A}{sample name, in column}	#'   \item{TCGA.60.2708.01A}{sample name, in column}	#'   \item{TCGA.60.2697.01A}{sample name, in column}	#'   \item{TCGA.63.A5MI.01A}{sample name, in column}	#'   \item{TCGA.56.7580.01A}{sample name, in column}	#'   \item{TCGA.43.A475.01A}{sample name, in column}	#'   \item{TCGA.77.8136.01A}{sample name, in column}	#'   \item{TCGA.21.A5DI.01A}{sample name, in column}	#'   \item{TCGA.85.8052.01A}{sample name, in column}	#'   \item{TCGA.21.1080.01A}{sample name, in column}	#'   \item{TCGA.77.A5FZ.01A}{sample name, in column}	#'   \item{TCGA.66.2737.01A}{sample name, in column}	#'   \item{TCGA.34.8456.01A}{sample name, in column}	#'   \item{TCGA.85.A4PA.01A}{sample name, in column}	#'   \item{TCGA.85.8355.01A}{sample name, in column}	#'   \item{TCGA.66.2742.01A}{sample name, in column}	#'   \item{TCGA.56.8626.01A}{sample name, in column}	#'   \item{TCGA.77.8140.01A}{sample name, in column}	#'   \item{TCGA.22.4601.01A}{sample name, in column}	#'   \item{TCGA.92.8064.01A}{sample name, in column}	#'   \item{TCGA.22.5481.01A}{sample name, in column}	#'   \item{TCGA.85.7950.01A}{sample name, in column}	#'   \item{TCGA.58.8386.01A}{sample name, in column}	#'   \item{TCGA.43.5668.01A}{sample name, in column}	#'   \item{TCGA.85.8350.01A}{sample name, in column}	#'   \item{TCGA.56.A49D.01A}{sample name, in column}	#'   \item{TCGA.21.5782.01A}{sample name, in column}	#'   \item{TCGA.77.8143.01A}{sample name, in column}	#'   \item{TCGA.85.7698.01A}{sample name, in column}	#'   \item{TCGA.85.8666.01A}{sample name, in column}	#'   \item{TCGA.56.8503.01A}{sample name, in column}	#'   \item{TCGA.34.5927.01A}{sample name, in column}	#'   \item{TCGA.85.A513.01A}{sample name, in column}	#'   \item{TCGA.NK.A5CX.01A}{sample name, in column}	#'   \item{TCGA.77.A5G8.01B}{sample name, in column}	#'   \item{TCGA.63.5131.01A}{sample name, in column}	#'   \item{TCGA.77.A5GH.01A}{sample name, in column}	#'   \item{TCGA.85.A4CL.01A}{sample name, in column}	#'   \item{TCGA.MF.A522.01A}{sample name, in column}	#'   \item{TCGA.66.2755.01A}{sample name, in column}	#'   \item{TCGA.77.7463.01A}{sample name, in column}	#'   \item{TCGA.68.7755.01A}{sample name, in column}	#'   \item{TCGA.NK.A5CT.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HI.01A}{sample name, in column}	#'   \item{TCGA.63.A5MJ.01A}{sample name, in column}	#'   \item{TCGA.77.A5G1.01A}{sample name, in column}	#'   \item{TCGA.18.3412.01A}{sample name, in column}	#'   \item{TCGA.39.5031.01A}{sample name, in column}	#'   \item{TCGA.22.5480.01A}{sample name, in column}	#'   \item{TCGA.94.8491.01A}{sample name, in column}	#'   \item{TCGA.85.6798.01A}{sample name, in column}	#'   \item{TCGA.85.8584.01A}{sample name, in column}	#'   \item{TCGA.66.2770.01A}{sample name, in column}	#'   \item{TCGA.85.6175.01A}{sample name, in column}	#'   \item{TCGA.60.2704.01A}{sample name, in column}	#'   \item{TCGA.60.2716.01A}{sample name, in column}	#'   \item{TCGA.21.1083.01A}{sample name, in column}	#'   \item{TCGA.56.8629.01A}{sample name, in column}	#'   \item{TCGA.43.8116.01A}{sample name, in column}	#'   \item{TCGA.33.4589.01A}{sample name, in column}	#'   \item{TCGA.63.A5MR.01A}{sample name, in column}	#'   \item{TCGA.90.A4ED.01A}{sample name, in column}	#'   \item{TCGA.63.5128.01A}{sample name, in column}	#'   \item{TCGA.O2.A52N.01A}{sample name, in column}	#'   \item{TCGA.33.A5GW.01A}{sample name, in column}	#'   \item{TCGA.60.2721.01A}{sample name, in column}	#'   \item{TCGA.90.A4EE.01A}{sample name, in column}	#'   \item{TCGA.22.5483.01A}{sample name, in column}	#'   \item{TCGA.68.A59I.01A}{sample name, in column}	#'   \item{TCGA.98.A53D.01A}{sample name, in column}	#'   \item{TCGA.52.7812.01A}{sample name, in column}	#'   \item{TCGA.39.5036.01A}{sample name, in column}	#'   \item{TCGA.63.A5MP.01A}{sample name, in column}	#'   \item{TCGA.77.A5G3.01A}{sample name, in column}	#'   \item{TCGA.77.8138.01A}{sample name, in column}	#'   \item{TCGA.46.6026.01A}{sample name, in column}	#'   \item{TCGA.85.7699.01A}{sample name, in column}	#'   \item{TCGA.85.8354.01A}{sample name, in column}	#'   \item{TCGA.85.8582.01A}{sample name, in column}	#'   \item{TCGA.39.5022.01A}{sample name, in column}	#'   \item{TCGA.34.2596.01A}{sample name, in column}	#'   \item{TCGA.XC.AA0X.01A}{sample name, in column}	#'   \item{TCGA.56.8623.01A}{sample name, in column}	#'   \item{TCGA.66.2777.01A}{sample name, in column}	#'   \item{TCGA.56.7823.01B}{sample name, in column}	#'   \item{TCGA.68.8251.01A}{sample name, in column}	#'   \item{TCGA.66.2734.01A}{sample name, in column}	#'   \item{TCGA.77.8153.01A}{sample name, in column}	#'   \item{TCGA.85.A4QR.01A}{sample name, in column}	#'   \item{TCGA.85.A4QQ.01A}{sample name, in column}	#'   \item{TCGA.22.4593.01A}{sample name, in column}	#'   \item{TCGA.66.2771.01A}{sample name, in column}	#'   \item{TCGA.58.8391.01A}{sample name, in column}	#'   \item{TCGA.77.8008.01A}{sample name, in column}	#'   \item{TCGA.60.2706.01A}{sample name, in column}	#'   \item{TCGA.34.8454.01A}{sample name, in column}	#'   \item{TCGA.L3.A524.01A}{sample name, in column}	#'   \item{TCGA.51.4079.01A}{sample name, in column}	#'   \item{TCGA.39.5021.01A}{sample name, in column}	#'   \item{TCGA.56.7222.01A}{sample name, in column}	#'   \item{TCGA.66.2783.01A}{sample name, in column}	#'   \item{TCGA.22.4596.01A}{sample name, in column}	#'   \item{TCGA.37.4133.01A}{sample name, in column}	#'   \item{TCGA.46.3768.01A}{sample name, in column}	#'   \item{TCGA.63.A5MS.01A}{sample name, in column}	#'   \item{TCGA.85.8288.01A}{sample name, in column}	#'   \item{TCGA.66.2794.01A}{sample name, in column}	#'   \item{TCGA.56.8307.01A}{sample name, in column}	#'   \item{TCGA.L3.A4E7.01A}{sample name, in column}	#'   \item{TCGA.56.5897.01A}{sample name, in column}	#'   \item{TCGA.58.8388.01A}{sample name, in column}	#'   \item{TCGA.60.2712.01A}{sample name, in column}	#'   \item{TCGA.34.5929.01A}{sample name, in column}	#'   \item{TCGA.34.A5IX.01A}{sample name, in column}	#'   \item{TCGA.56.8083.01A}{sample name, in column}	#'   \item{TCGA.33.AAS8.01A}{sample name, in column}	#'   \item{TCGA.18.3410.01A}{sample name, in column}	#'   \item{TCGA.60.2713.01A}{sample name, in column}	#'   \item{TCGA.66.2789.01A}{sample name, in column}	#'   \item{TCGA.34.5234.01A}{sample name, in column}	#'   \item{TCGA.37.A5EL.01A}{sample name, in column}	#'   \item{TCGA.O2.A5IB.01A}{sample name, in column}	#'   \item{TCGA.43.A56U.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HF.01A}{sample name, in column}	#'   \item{TCGA.37.A5EN.01A}{sample name, in column}	#'   \item{TCGA.56.A4BW.01A}{sample name, in column}	#'   \item{TCGA.66.2753.01A}{sample name, in column}	#'   \item{TCGA.85.8664.01A}{sample name, in column}	#'   \item{TCGA.77.8156.01A}{sample name, in column}	#'   \item{TCGA.34.8455.01A}{sample name, in column}	#'   \item{TCGA.63.A5M9.01A}{sample name, in column}	#'   \item{TCGA.66.2759.01A}{sample name, in column}	#'   \item{TCGA.56.7582.01A}{sample name, in column}	#'   \item{TCGA.96.7545.01A}{sample name, in column}	#'   \item{TCGA.85.6561.01A}{sample name, in column}	#'   \item{TCGA.6A.AB49.01A}{sample name, in column}	#'   \item{TCGA.85.7696.01A}{sample name, in column}	#'   \item{TCGA.43.3920.01A}{sample name, in column}	#'   \item{TCGA.56.8305.01A}{sample name, in column}	#'   \item{TCGA.58.A46J.01A}{sample name, in column}	#'   \item{TCGA.46.3769.01A}{sample name, in column}	#'   \item{TCGA.NK.A5D1.01A}{sample name, in column}	#'   \item{TCGA.18.3415.01A}{sample name, in column}	#'   \item{TCGA.63.A5MT.01A}{sample name, in column}	#'   \item{TCGA.66.2795.01A}{sample name, in column}	#'   \item{TCGA.37.4135.01A}{sample name, in column}	#'   \item{TCGA.90.6837.01A}{sample name, in column}	#'   \item{TCGA.63.A5MV.01A}{sample name, in column}	#'   \item{TCGA.85.8352.01A}{sample name, in column}	#'   \item{TCGA.77.A5GB.01B}{sample name, in column}	#'   \item{TCGA.70.6723.01A}{sample name, in column}	#'   \item{TCGA.85.7697.01A}{sample name, in column}	#'   \item{TCGA.63.A5MM.01A}{sample name, in column}	#'   \item{TCGA.43.A474.01A}{sample name, in column}	#'   \item{TCGA.66.2754.01A}{sample name, in column}	#'   \item{TCGA.56.8622.01A}{sample name, in column}	#'   \item{TCGA.18.3406.01A}{sample name, in column}	#'   \item{TCGA.22.5485.01A}{sample name, in column}	#'   \item{TCGA.66.2767.01A}{sample name, in column}	#'   \item{TCGA.85.7843.01A}{sample name, in column}	#'   \item{TCGA.85.8351.01A}{sample name, in column}	#'   \item{TCGA.34.5241.01A}{sample name, in column}	#'   \item{TCGA.39.5039.01A}{sample name, in column}	#'   \item{TCGA.66.2768.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HN.01A}{sample name, in column}	#'   \item{TCGA.85.7844.01A}{sample name, in column}	#'   \item{TCGA.98.A53J.01A}{sample name, in column}	#'   \item{TCGA.46.6025.01A}{sample name, in column}	#'   \item{TCGA.56.6545.01A}{sample name, in column}	#'   \item{TCGA.85.A4CN.01A}{sample name, in column}	#'   \item{TCGA.63.A5ML.01A}{sample name, in column}	#'   \item{TCGA.63.7021.01A}{sample name, in column}	#'   \item{TCGA.60.2722.01A}{sample name, in column}	#'   \item{TCGA.98.A53H.01A}{sample name, in column}	#'   \item{TCGA.33.4587.01A}{sample name, in column}	#'   \item{TCGA.56.8504.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HJ.01A}{sample name, in column}	#'   \item{TCGA.46.3767.01A}{sample name, in column}	#'   \item{TCGA.43.7657.01A}{sample name, in column}	#'   \item{TCGA.21.1077.01A}{sample name, in column}	#'   \item{TCGA.85.A4JB.01A}{sample name, in column}	#'   \item{TCGA.51.4080.01A}{sample name, in column}	#'   \item{TCGA.85.8580.01A}{sample name, in column}	#'   \item{TCGA.85.8479.01A}{sample name, in column}	#'   \item{TCGA.18.3419.01A}{sample name, in column}	#'   \item{TCGA.22.5492.01A}{sample name, in column}	#'   \item{TCGA.92.8063.01A}{sample name, in column}	#'   \item{TCGA.77.8130.01A}{sample name, in column}	#'   \item{TCGA.34.5239.01A}{sample name, in column}	#'   \item{TCGA.66.2792.01A}{sample name, in column}	#'   \item{TCGA.90.7767.01A}{sample name, in column}	#'   \item{TCGA.98.8020.01A}{sample name, in column}	#'   \item{TCGA.37.4141.01A}{sample name, in column}	#'   \item{TCGA.33.AASL.01A}{sample name, in column}	#'   \item{TCGA.85.A511.01A}{sample name, in column}	#'   \item{TCGA.58.8392.01A}{sample name, in column}	#'   \item{TCGA.77.8131.01A}{sample name, in column}	#'   \item{TCGA.22.5477.01A}{sample name, in column}	#'   \item{TCGA.43.6773.01A}{sample name, in column}	#'   \item{TCGA.68.7756.01A}{sample name, in column}	#'   \item{TCGA.43.3394.01A}{sample name, in column}	#'   \item{TCGA.94.A4VJ.01A}{sample name, in column}	#'   \item{TCGA.66.2727.01A}{sample name, in column}	#'   \item{TCGA.22.4607.01A}{sample name, in column}	#'   \item{TCGA.56.1622.01A}{sample name, in column}	#'   \item{TCGA.66.2781.01A}{sample name, in column}	#'   \item{TCGA.77.7465.01A}{sample name, in column}	#'   \item{TCGA.37.4130.01A}{sample name, in column}	#'   \item{TCGA.77.7337.01A}{sample name, in column}	#'   \item{TCGA.21.1076.01A}{sample name, in column}	#'   \item{TCGA.77.A5GA.01A}{sample name, in column}	#'   \item{TCGA.34.5231.01A}{sample name, in column}	#'   \item{TCGA.60.2723.01A}{sample name, in column}	#'   \item{TCGA.22.5489.01A}{sample name, in column}	#'   \item{TCGA.52.7622.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HD.01A}{sample name, in column}	#'   \item{TCGA.21.1072.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HQ.01A}{sample name, in column}	#'   \item{TCGA.63.A5MU.01A}{sample name, in column}	#'   \item{TCGA.92.7341.01A}{sample name, in column}	#'   \item{TCGA.56.8309.01A}{sample name, in column}	#'   \item{TCGA.21.5786.01A}{sample name, in column}	#'   \item{TCGA.56.5898.01A}{sample name, in column}	#'   \item{TCGA.77.7338.01A}{sample name, in column}	#'   \item{TCGA.94.8035.01A}{sample name, in column}	#'   \item{TCGA.56.A5DR.01A}{sample name, in column}	#'   \item{TCGA.22.5471.01A}{sample name, in column}	#'   \item{TCGA.52.7809.01A}{sample name, in column}	#'   \item{TCGA.96.8170.01A}{sample name, in column}	#'   \item{TCGA.56.8304.01A}{sample name, in column}	#'   \item{TCGA.18.4721.01A}{sample name, in column}	#'   \item{TCGA.77.8009.01A}{sample name, in column}	#'   \item{TCGA.33.6738.01A}{sample name, in column}	#'   \item{TCGA.51.4081.01A}{sample name, in column}	#'   \item{TCGA.43.5670.01A}{sample name, in column}	#'   \item{TCGA.O2.A52V.01A}{sample name, in column}	#'   \item{TCGA.22.5478.01A}{sample name, in column}	#'   \item{TCGA.90.A59Q.01A}{sample name, in column}	#'   \item{TCGA.34.5928.01A}{sample name, in column}	#'   \item{TCGA.46.3765.01A}{sample name, in column}	#'   \item{TCGA.58.8387.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HL.01A}{sample name, in column}	#'   \item{TCGA.22.5473.01A}{sample name, in column}	#'   \item{TCGA.34.5240.01A}{sample name, in column}	#'   \item{TCGA.85.6560.01A}{sample name, in column}	#'   \item{TCGA.NK.A5CR.01A}{sample name, in column}	#'   \item{TCGA.37.5819.01A}{sample name, in column}	#'   \item{TCGA.39.5035.01A}{sample name, in column}	#'   \item{TCGA.85.A512.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HH.01A}{sample name, in column}	#'   \item{TCGA.37.3783.01A}{sample name, in column}	#'   \item{TCGA.94.A5I6.01A}{sample name, in column}	#'   \item{TCGA.18.5595.01A}{sample name, in column}	#'   \item{TCGA.56.A4ZK.01A}{sample name, in column}	#'   \item{TCGA.18.3421.01A}{sample name, in column}	#'   \item{TCGA.LA.A7SW.01A}{sample name, in column}	#'   \item{TCGA.60.2711.01A}{sample name, in column}	#'   \item{TCGA.33.AASI.01A}{sample name, in column}	#'   \item{TCGA.56.8082.01A}{sample name, in column}	#'   \item{TCGA.60.2710.01A}{sample name, in column}	#'   \item{TCGA.33.4547.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HT.01A}{sample name, in column}	#'   \item{TCGA.60.2696.01A}{sample name, in column}	#'   \item{TCGA.33.4532.01A}{sample name, in column}	#'   \item{TCGA.68.8250.01A}{sample name, in column}	#'   \item{TCGA.63.A5MW.01A}{sample name, in column}	#'   \item{TCGA.77.6844.01A}{sample name, in column}	#'   \item{TCGA.56.7731.01A}{sample name, in column}	#'   \item{TCGA.98.A539.01A}{sample name, in column}	#'   \item{TCGA.22.4613.01A}{sample name, in column}	#'   \item{TCGA.92.7340.01A}{sample name, in column}	#'   \item{TCGA.39.5028.01A}{sample name, in column}	#'   \item{TCGA.43.6647.01A}{sample name, in column}	#'   \item{TCGA.63.7022.01A}{sample name, in column}	#'   \item{TCGA.66.2758.01A}{sample name, in column}	#'   \item{TCGA.60.2724.01A}{sample name, in column}	#'   \item{TCGA.77.8145.01A}{sample name, in column}	#'   \item{TCGA.58.A46K.01A}{sample name, in column}	#'   \item{TCGA.94.7033.01A}{sample name, in column}	#'   \item{TCGA.43.2576.01A}{sample name, in column}	#'   \item{TCGA.85.8481.01A}{sample name, in column}	#'   \item{TCGA.39.5030.01A}{sample name, in column}	#'   \item{TCGA.33.6737.01A}{sample name, in column}	#'   \item{TCGA.77.7141.01A}{sample name, in column}	#'   \item{TCGA.18.3414.01A}{sample name, in column}	#'   \item{TCGA.77.7138.01A}{sample name, in column}	#'   \item{TCGA.33.AASD.01A}{sample name, in column}	#'   \item{TCGA.34.5232.01A}{sample name, in column}	#'   \item{TCGA.85.A510.01A}{sample name, in column}	#'   \item{TCGA.39.5034.01A}{sample name, in column}	#'   \item{TCGA.22.1012.01A}{sample name, in column}	#'   \item{TCGA.98.A53A.01A}{sample name, in column}	#'   \item{TCGA.58.8390.01A}{sample name, in column}	#'   \item{TCGA.22.4604.01A}{sample name, in column}	#'   \item{TCGA.77.8146.01A}{sample name, in column}	#'   \item{TCGA.56.7730.01A}{sample name, in column}	#'   \item{TCGA.18.5592.01A}{sample name, in column}	#'   \item{TCGA.33.4538.01A}{sample name, in column}	#'   \item{TCGA.58.A46M.01A}{sample name, in column}	#'   \item{TCGA.43.6770.01A}{sample name, in column}	#'   \item{TCGA.33.A4WN.01A}{sample name, in column}	#'   \item{TCGA.O2.A52Q.01A}{sample name, in column}	#'   \item{TCGA.22.5491.01A}{sample name, in column}	#'   \item{TCGA.85.8048.01A}{sample name, in column}	#'   \item{TCGA.22.5479.01A}{sample name, in column}	#'   \item{TCGA.37.3792.01A}{sample name, in column}	#'   \item{TCGA.63.A5MH.01A}{sample name, in column}	#'   \item{TCGA.33.AASB.01A}{sample name, in column}	#'   \item{TCGA.63.7020.01A}{sample name, in column}	#'   \item{TCGA.70.6722.01A}{sample name, in column}	#'   \item{TCGA.56.8628.01A}{sample name, in column}	#'   \item{TCGA.22.0940.01A}{sample name, in column}	#'   \item{TCGA.37.A5EM.01A}{sample name, in column}	#'   \item{TCGA.63.A5MB.01A}{sample name, in column}	#'   \item{TCGA.98.A53B.01A}{sample name, in column}	#'   \item{TCGA.94.7557.01A}{sample name, in column}	#'   \item{TCGA.58.A46L.01A}{sample name, in column}	#'   \item{TCGA.52.7810.01A}{sample name, in column}	#'   \item{TCGA.34.5236.01A}{sample name, in column}	#'   \item{TCGA.21.1075.01A}{sample name, in column}	#'   \item{TCGA.98.A538.01A}{sample name, in column}	#'   \item{TCGA.56.8308.01A}{sample name, in column}	#'   \item{TCGA.90.7766.01A}{sample name, in column}	#'   \item{TCGA.68.7757.01B}{sample name, in column}	#'   \item{TCGA.77.8144.01A}{sample name, in column}	#'   \item{TCGA.85.8287.01A}{sample name, in column}	#'   \item{TCGA.96.7544.01A}{sample name, in column}	#'   \item{TCGA.98.8023.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HE.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HO.01A}{sample name, in column}	#'   \item{TCGA.85.7710.01A}{sample name, in column}	#'   \item{TCGA.43.7658.01A}{sample name, in column}	#'   \item{TCGA.77.8150.01A}{sample name, in column}	#'   \item{TCGA.77.7335.01A}{sample name, in column}	#'   \item{TCGA.56.8201.01A}{sample name, in column}	#'   \item{TCGA.60.2695.01A}{sample name, in column}	#'   \item{TCGA.66.2744.01A}{sample name, in column}	#'   \item{TCGA.85.A4JC.01A}{sample name, in column}	#'   \item{TCGA.77.8128.01A}{sample name, in column}	#'   \item{TCGA.60.2698.01A}{sample name, in column}	#'   \item{TCGA.85.A53L.01A}{sample name, in column}	#'   \item{TCGA.85.8353.01A}{sample name, in column}	#'   \item{TCGA.43.A56V.01A}{sample name, in column}	#'   \item{TCGA.56.A4ZJ.01A}{sample name, in column}	#'   \item{TCGA.NC.A5HP.01A}{sample name, in column}	#'   \item{TCGA.66.2780.01A}{sample name, in column}	#'   \item{TCGA.92.8065.01A}{sample name, in column}	#'   \item{TCGA.96.A4JK.01A}{sample name, in column}	#'   \item{TCGA.66.2785.01A}{sample name, in column}	#'   \item{TCGA.77.A5G7.01B}{sample name, in column}	#'   \item{TCGA.39.5019.01A}{sample name, in column}	#'   \item{TCGA.85.8072.01A}{sample name, in column}	#'   \item{TCGA.52.7811.01A}{sample name, in column}	#'   \item{TCGA.22.1011.01A}{sample name, in column}	#'   \item{TCGA.O2.A52W.01A}{sample name, in column}	#'   \item{TCGA.63.A5MN.01A}{sample name, in column}	#'   \item{TCGA.39.5029.01A}{sample name, in column}	#'   \item{TCGA.66.2800.01A}{sample name, in column}	#'   \item{TCGA.66.2782.01A}{sample name, in column}	#'   \item{TCGA.22.0944.01A}{sample name, in column}	#'   \item{TCGA.85.A50M.01A}{sample name, in column}	#'   \item{TCGA.90.7769.01A}{sample name, in column}	#'   \item{TCGA.43.6143.01A}{sample name, in column}	#'   \item{TCGA.98.7454.01A}{sample name, in column}	#'   \item{TCGA.56.8624.01A}{sample name, in column}	#'   \item{TCGA.51.6867.01A}{sample name, in column}	#'   \item{TCGA.85.8049.01A}{sample name, in column}	#'   \item{TCGA.66.2788.01A}{sample name, in column}	#'   \item{TCGA.43.2581.01A}{sample name, in column}	#'   \item{TCGA.O2.A52S.01A}{sample name, in column}	#'   \item{TCGA.39.5027.01A}{sample name, in column}	#'   \item{TCGA.77.8148.01A}{sample name, in column}	#'   \item{TCGA.60.2703.01A}{sample name, in column}	#'   \item{TCGA.22.A5C4.01A}{sample name, in column}	#'   \item{TCGA.79.5596.01A}{sample name, in column}	#'   \item{TCGA.85.8070.01A}{sample name, in column}	#'   \item{TCGA.21.1079.01A}{sample name, in column}	#'   \item{TCGA.96.8169.01A}{sample name, in column}	#'   \item{TCGA.J1.A4AH.01A}{sample name, in column}	#'   \item{TCGA.77.6842.01A}{sample name, in column}	#'   \item{TCGA.43.8115.01A}{sample name, in column}	#'   \item{TCGA.66.2763.01A}{sample name, in column}	#'   \item{TCGA.43.8118.01A}{sample name, in column}	#'   \item{TCGA.85.8071.01A}{sample name, in column}	#'   \item{TCGA.85.A5B5.01A}{sample name, in column}
  }}
\source{
TCGA lung adenocarcinoma:
 \url{https://tcga-data.nci.nih.gov/docs/publications/tcga/}
}
\usage{
gene_sam
}
\value{
data.frame
}
\description{
This gene expression dataset is originally from TCGA lung adenocarcinoma. To make dataset
smaller, we have selected 475 samples in columns and  918 genes in rows.
}
\keyword{datasets}
