% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRhubgene.R
\name{miRhubgene}
\alias{miRhubgene}
\title{sort the information to find out the triplets.}
\usage{
miRhubgene(dictionary, Result_TCGA_LUSC)
}
\arguments{
\item{dictionary}{miRNA and the corresponding genes' combination in list format, with miRNA name in the first column and the corresponding genes in the second column}

\item{Result_TCGA_LUSC}{a list of dataframe formats with miRNA,candidate ceRNAs,and their locations and number of segmentation based on segcluster_peakmerge function.}
}
\value{
outputm a dataframe formats with miRNA names with the number of bridging ceRNA triplets and the corresponding genes with the number of ceRNA triplets.
}
\description{
This function will sort the information of miRNA and ceRNA to find out triplets.
}
\examples{
## Use the internal dataset

data("dictionary", package = "ceRNAmiRNAfun", envir = environment())
data("Result_TCGA_LUSC", package = "ceRNAmiRNAfun", envir = environment())

## use miRhubgene to merge out the bridging miRNA and hub genes
miRhubgene(dictionary,Result_TCGA_LUSC)
  

}
