% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRhubgeneplot.R
\name{miRhubgeneplot}
\alias{miRhubgeneplot}
\title{sort the information to find out the triplets and plot the top miRNA expression level of modulated ceRNA interaction happened}
\usage{
miRhubgeneplot(dictionary, mirna_sam, Result_TCGA_LUSC, miRhubgeneoutput,
  w, N)
}
\arguments{
\item{dictionary}{miRNA and the corresponding genes' combination in list format, with miRNA name in the first column and the corresponding genes in the second column}

\item{mirna_sam}{expression data of miRNA in dataframe format, with miRNA's name in rows and sample name in columns}

\item{Result_TCGA_LUSC}{a list of dataframe formats with miRNA,candidate ceRNAs,and their locations and number of segmentation based on segcluster_peakmerge function.}

\item{miRhubgeneoutput}{a dataframe formats with miRNA names with the number of bridging ceRNA triplets and the corresponding genes with the number of ceRNA triplets.}

\item{w}{window size of the each calculation of triplets,the default is 10.}

\item{N}{number of the samples,the default is 475.}
}
\value{
plotp a plot format with miRNA expression in x-axis and the interaction ceRNA in y-axis.
}
\description{
This function will sort the information to find out the triplets and plot the top miRNA expression level of modulated ceRNA interaction happened
}
\examples{
## Use the internal dataset

data("dictionary", package = "ceRNAmiRNAfun", envir = environment())
data("mirna_sam", package = "ceRNAmiRNAfun", envir = environment())
data("Result_TCGA_LUSC", package = "ceRNAmiRNAfun", envir = environment())
data("miRhubgeneoutput", package = "ceRNAmiRNAfun", envir = environment())

## use miRhubgeneplot to merge out the bridging miRNA and hub genes and then make the plot of the top miRNA expression level of modulated ceRNA interaction happened
miRhubgeneplot(dictionary,mirna_sam,Result_TCGA_LUSC,miRhubgeneoutput,w=10,N=475)
  
}
