% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_miRNA_total.R
\docType{data}
\name{miRNA_total}
\alias{miRNA_total}
\title{A list with information of miRNAs}
\format{A vector with 137miRNA names}
\source{
TCGA lung adenocarcinoma:
 \url{https://tcga-data.nci.nih.gov/docs/publications/tcga/}
}
\usage{
miRNA_total
}
\value{
vector
}
\description{
This miRNA dataset is originally TCGA lung adenocarcinoma. It contains
137 miRNA names
}
\keyword{datasets}
