% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{Random_walk}
\alias{Random_walk}
\title{evaluate the correlation coefficients between two genes to find out the pair ceRNAs of a miRNA.}
\usage{
Random_walk(triplet, cor_shreshold_rw)
}
\arguments{
\item{triplet}{two genes and a miRNA combination.}

\item{cor_shreshold_rw}{the shreshold of correlation of coefficient between two genes.}
}
\value{
numeric format with P value of each pair of genes with each miRNA in overall triplet.
}
\description{
This function will evaluate the fraction of samples which their gene correlations are under the condition weighted by their correlation and the fractions that are not presnt up to a given position in all samples. And then, we sort the genes with increasing expression level of a miRNA, by permuting the miRNA expression for 1000 times, and computed S, the maximum deviation from zero of the fractions of samples which their gene correlations are under the condition and those are not, of the permuted data. The empirical P value is calculated relative to this null distribution. Based on the P value, we can find out the ceRNAs of a miRNA.
}
\examples{
## Use the internal dataset
data("triplet", package = "anamiR", envir = environment())

## evaluate correlation coefficents between two genes.
Random_walk(triplet,cor_shreshold_rw=0.5)

}
