% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Realdata.R
\docType{data}
\name{Realdata}
\alias{Realdata}
\title{a list of dataframe formats with correlation
coefficients of genes within each window for 
each corresponding miRNA average expression.}
\format{A large dataframe with miRNA's expression in 
the first column, and the correlation coefficient in 
the folloing columns}
\source{
TCGA data portal:
 \url{https://tcga-data.nci.nih.gov/docs/publications/tcga/}
}
\usage{
Realdata
}
\value{
list
}
\description{
This dataframe is generated from \link{sliding_windows},
}
\keyword{datasets}
