% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Result_TCGA_LUSC.R
\docType{data}
\name{Result_TCGA_LUSC}
\alias{Result_TCGA_LUSC}
\title{a list of dataframe formats with miRNA,
candidate ceRNAs,and their locations and  
number of segmentation based on 
segcluster_peakmerge function.}
\format{A large dataframe with miRNA's name in 
the first column, candidate ceRNA names, the ceRNA's
locations, and the number of segmentation as order in
in the following columns.}
\source{
TCGA data portal:
 \url{https://tcga-data.nci.nih.gov/docs/publications/tcga/}
}
\usage{
Result_TCGA_LUSC
}
\value{
dataframe
}
\description{
This dataframe is generated from function segcluster_peakmerge.
}
\keyword{datasets}
