% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_windows.R
\name{Datalist_sliding_w}
\alias{Datalist_sliding_w}
\title{calculate correlation coefficients with ceRNAs of a miRNA by sliding windows.}
\usage{
Datalist_sliding_w(w, miRNA_total)
}
\arguments{
\item{w}{window size of the each calculation of triplets,the default is 105.}

\item{miRNA_total}{miRNA dataset in a dataframe format}
}
\value{
dataframe format with correlation coefficients within each window for each corresponding miRNA average expression.
}
\description{
This function will calculate correlation coefficients within each window, a sliding mover that contains putative ceRNA triplets composed of a miRNA and several genes, after determing a window size and the number of windows. And the corresponding miRNA expression of each window was the average expression of samples within different windows.
}
\examples{
## Use the internal dataset
data("mirna_sam", package = "anamiR", envir = environment())
data("gene_sam", package = "anamiR", envir = environment())


## use sliding window to calculate correlations
Datalist_sliding_w(w=105,miRNA_total=mirna_total)
  

}
